+++ R/flag-dupes.R
@@ -15,7 +15,7 @@
 #' @export
 flag_dupes <- function(data, ..., .check = TRUE) {
   sub_data <- dplyr::select(data, ...)
-  dupe_vec <- duplicated(sub_data)
+  dupe_vec <- duplicated(sub_data) | duplicated(sub_data, fromLast = TRUE)
   if (.check & sum(dupe_vec) == 0) {
     warning("no duplicate rows, column not created")
     return(data)
@@ -23,3 +23,7 @@ flag_dupes <- function(data, ..., .check = TRUE) {
     dplyr::mutate(data, dupe_flag = dupe_vec)
   }
 }
+
+sub_data <- dplyr::select(iris, everything())
+dupe_vec <- duplicated(sub_data)
+dplyr::mutate(sub_data, dupe_flag = dupe_vec)